requirements.txt
psycopg2-binary>=2.9
pgvector>=0.3
requests>=2.31
python-dotenv>=1.0
rich>=13.0
numpy>=1.26
